Move Statistics from Settings to a sidebar Tools view
Statistics is a read-only network monitor, not configuration, so it now
lives in the main window as a sidebar tool (#statistics) rendered by
ConversationPane like the other tool views. The old #settings/statistics
hash redirects to the new route, and last-viewed restore, the command
palette, and lazy loading all treat it like the existing tools.

Also tightens isMessageConversation to an allowlist (channel/contact) so
non-chat views like bots and statistics never trigger message loading.
